Force var function to return a vector
When data is passed in where a batch only has a single sample
corresponding to it, `var(s_data(:, ind1)` returns a scalar, which
causes the cat command to fail. By passing in the dimension parameter
to the `var` command, it forces `var` to return a vector and prevents
this problem. We can then transpose the output of `var` to keep the
resulting dimensions the same as they were before.